+++ mde-services/src/main/java/de/terrestris/mde/mde_backend/model/json/JsonClientMetadata.java
@@ -1,20 +1,21 @@
 package de.terrestris.mde.mde_backend.model.json;
 
 import com.fasterxml.jackson.annotation.JsonInclude;
-import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
+import java.io.Serializable;
 import java.util.List;
 import java.util.Map;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
-import org.springframework.lang.Nullable;
+import org.jspecify.annotations.Nullable;
+import tools.jackson.databind.annotation.JsonDeserialize;
 
 @Data
 @JsonDeserialize(as = JsonClientMetadata.class)
 @JsonInclude(JsonInclude.Include.NON_NULL)
 @NoArgsConstructor(force = true)
 @AllArgsConstructor
-public class JsonClientMetadata {
+public class JsonClientMetadata implements Serializable {
 
   @Nullable private List<Comment> comments;