+++ mde-services/src/main/java/de/terrestris/mde/mde_backend/model/json/JsonTechnicalMetadata.java
@@ -2,18 +2,19 @@ package de.terrestris.mde.mde_backend.model.json;
 
 import com.fasterxml.jackson.annotation.JsonInclude;
 import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
+import java.io.Serializable;
 import java.util.List;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
-import org.springframework.lang.Nullable;
+import org.jspecify.annotations.Nullable;
 
 @Data
 @JsonDeserialize(as = JsonTechnicalMetadata.class)
 @JsonInclude(JsonInclude.Include.NON_NULL)
 @NoArgsConstructor(force = true)
 @AllArgsConstructor
-public class JsonTechnicalMetadata {
+public class JsonTechnicalMetadata implements Serializable {
 
   @Nullable private List<LayerInfo> layerInfos;