+++ mde-services/src/main/java/de/terrestris/mde/mde_backend/service/ValidatorService.java
@@ -3,11 +3,9 @@ package de.terrestris.mde.mde_backend.service;
 import static de.terrestris.bkgtestsuite.core.model.TestResult.MessageType.TEXT;
 import static de.terrestris.bkgtestsuite.core.model.TestResult.ResultStatus.*;
 
-import com.fasterxml.jackson.databind.ObjectMapper;
 import de.terrestris.bkgtestsuite.core.model.TestParameters;
 import de.terrestris.bkgtestsuite.core.model.TestResult;
 import de.terrestris.bkgtestsuite.core.spi.TestRunner;
-import de.terrestris.bkgtestsuite.etf.EtfTestRunner;
 import de.terrestris.bkgtestsuite.te.TeTestRunner;
 import de.terrestris.mde.mde_backend.enumeration.MetadataProfile;
 import de.terrestris.mde.mde_backend.enumeration.ValidationStatus;
@@ -20,7 +18,6 @@ import java.io.File;
 import java.io.IOException;
 import java.nio.file.Files;
 import java.nio.file.Path;
-import java.security.NoSuchAlgorithmException;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Map;
@@ -38,6 +35,7 @@ import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.security.core.context.SecurityContext;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Component;
+import tools.jackson.databind.ObjectMapper;
 
 @Component
 @Log4j2
@@ -64,12 +62,6 @@ public class ValidatorService {
 
   private TestRunner inspireRunner;
 
-  private TestRunner etfServiceRunner;
-
-  private TestRunner etfDatasetRunner;
-
-  private TestRunner etfNetworkServicesRunner;
-
   private final List<Map<String, String>> inspireClasses =
       List.of(
           Map.of("id", "ISO-Schemavalidierung"),
@@ -86,29 +78,6 @@ public class ValidatorService {
           Map.of("id", "GDI-DE_konditional"),
           Map.of("id", "GDI-DE_optional"),
           Map.of("id", "OpenData_konditional"));
-
-  private final List<Map<String, String>> datasetClasses =
-      List.of(
-          Map.of("id", "EID59692c11-df86-49ad-be7f-94a1e1ddd8da"),
-          Map.of("id", "EIDe4a95862-9cc9-436b-9fdd-a0115d342350"),
-          Map.of("id", "EID2be1480a-fe42-40b2-9420-eb0e69385c80"),
-          Map.of("id", "EID0b86f7a3-2947-4841-823d-6a00d8e06d70"),
-          Map.of("id", "EID1067d6b2-3bb1-4e71-8ce1-a744c9bd5a3b"),
-          Map.of("id", "EID7cceba68-e575-4429-9959-1b6b3d201b6d"));
-
-  private final List<Map<String, String>> serviceClasses =
-      List.of(
-          Map.of("id", "EID59692c11-df86-49ad-be7f-94a1e1ddd8da"),
-          Map.of("id", "EID8f869e23-c9e9-4e86-8dca-be30ff421229"),
-          Map.of("id", "EID8db54d8a-8578-4959-b891-5394d9f53a28"),
-          Map.of("id", "EID7514777a-6cb8-499c-acd5-912496dc84e9"),
-          Map.of("id", "EIDa593a7ad-42d9-46d0-985d-9dff3e684428"));
-
-  private final List<Map<String, String>> networkServiceClasses =
-      List.of(
-          Map.of("id", "EID59692c11-df86-49ad-be7f-94a1e1ddd8da"),
-          Map.of("id", "EID8f869e23-c9e9-4e86-8dca-be30ff421229"),
-          Map.of("id", "EID606587df-65a8-4b7b-9eee-e0d94daaa42a"));
 
   @PostConstruct
   public void loadIsoPackage() throws IOException {
@@ -130,24 +99,6 @@ public class ValidatorService {
             inspireConfig.getInputStream(),
             list,
             new File("/tmp/bkg-scripts-metadata/src/resources"));
-  }
-
-  @PostConstruct
-  public void loadEtfServicePackage() throws IOException, NoSuchAlgorithmException {
-    var node = MAPPER.readTree(getClass().getResourceAsStream("/config-services.json"));
-    etfServiceRunner = new EtfTestRunner(node, false);
-  }
-
-  @PostConstruct
-  public void loadEtfDatasetPackage() throws IOException, NoSuchAlgorithmException {
-    var node = MAPPER.readTree(getClass().getResourceAsStream("/config-dataset.json"));
-    etfDatasetRunner = new EtfTestRunner(node, false);
-  }
-
-  @PostConstruct
-  public void loadEtfNetworkServicesPackage() throws IOException, NoSuchAlgorithmException {
-    var node = MAPPER.readTree(getClass().getResourceAsStream("/config-networkservices.json"));
-    etfNetworkServicesRunner = new EtfTestRunner(node, false);
   }
 
   private static List<File> extractResources(List<Resource> resources, Predicate<File> tester)
@@ -195,22 +146,6 @@ public class ValidatorService {
           errors,
           classes,
           inspire ? "Ergebnisse GDI-DE für INSPIRE:" : "Ergebnisse GDI-DE:");
-      if (inspire) {
-        if (dataset) {
-          log.debug("Running INSPIRE dataset tests...");
-          runTest(file, etfDatasetRunner, errors, datasetClasses, "Ergebnisse INSPIRE-Validator:");
-        } else {
-          log.debug("Running INSPIRE service tests...");
-          runTest(file, etfServiceRunner, errors, serviceClasses, "Ergebnisse INSPIRE-Validator:");
-          log.debug("Running INSPIRE network services tests...");
-          runTest(
-              file,
-              etfNetworkServicesRunner,
-              errors,
-              networkServiceClasses,
-              "Ergebnisse INSPIRE-Validator:");
-        }
-      }
       return errors;
     } catch (Exception e) {
       log.error("Error when validating: {}", e.getMessage());